renamed book containers to BookOverview & BookDetails
